104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,15 +7,15 @@
  */
 int main(void)
 {
-	unsigned long int prev1 = 1, prev2 = 2, next;
+	unsigned long long int prev1 = 1, prev2 = 2, next;
 	int i;
 
-	printf("%lu, %lu", prev1, prev2);
+	printf("%llu, %llu", prev1, prev2);
 
 	for (i = 2; i < 98; i++)
 	{
 		next = prev1 + prev2;
-		printf(", %lu", next);
+		printf(", %llu", next);
 		prev1 = prev2;
 		prev2 = next;
 	}
